Fixed Paging Bug
The way request URLs were being built was the individual client
methods would call the generic method with an endpoint. The
generic endpoint would then concatenate the hostname URL with
the endpoint and use that. When next page requests would come
in the method would append the full URL to the base.

This moves the URL building to the methods and the generic request
method now naively uses the URL passed to it.

Big thanks to @seanshahkarami for posting the issue!

Updates #3