Added Metrics Endpoint
I recently added a new `metrics` endpoint to the API to support
applications that read the nodes' telemetry data. The new endpoint is a
copy of what the observations endpoint is -- the only difference is
segmenting the results by the sensor path.

This adds a new `list_metrics` method to the client to access that
endpoint.